Implemented TermList class (non-functional)

+++ src/algorithms/Term.java
@@ -2,11 +2,11 @@ package algorithms;
 
 public class Term {
 	
-	public int weight;
+	public double weight;
 	public String text;
 	
 	
-	public Term (int weight, String text)
+	public Term (double weight, String text)
 	{
 		this.weight = weight;
 		this.text = text;
